팔로우기능구현

+++ src/main/java/com/mwom/moyeora/database/entity/MemberInfoEntity.java
@@ -25,11 +25,11 @@ public class MemberInfoEntity {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)//프로젝트에서 연결된 DB의 넘버링 전략을 따라간다.
   @Column(name = "MEMBER_INFO_SEQ")
-  private int memberInfoSeq;
+  private long memberInfoSeq;
 
 
   @Column(name = "MEMBER_SEQ")
-  private int memberSeq;
+  private long memberSeq;
 
   @Column(name = "ADDRESS_NUM")
   private String addressNum;
